+++ Assets/Scripts/Zones/Transitions/Fader.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections;
 using UnityEngine;
 using UnityEngine.UI;
-using Frankie.Core;
 using Frankie.Rendering;
+using Frankie.Saving;
 
 namespace Frankie.ZoneManagement
 {
